# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Transaction API written in Go",
        "title": "Transaction API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/transactions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get All Transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetAllTransactionsResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Update Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UpdateByIdTransactionResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create Transaction",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CreateTransactionResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Delete Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/DeleteByIdTransactionResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get One Transaction By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetOneTransactionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateTransactionRequest": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string",
                    "example": "233455011"
                },
                "amount": {
                    "type": "number",
                    "example": 10.3
                },
                "transaction_type": {
                    "type": "string",
                    "example": "credit"
                }
            }
        },
        "CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/TransactionResponse"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "DeleteByIdTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DeleteTransactionResponse"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "DeleteTransactionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "data deleted successfully"
                }
            }
        },
        "GetAllTransactionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TransactionResponse"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "GetOneTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/TransactionResponse"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "TransactionResponse": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string",
                    "example": "233455011"
                },
                "amount": {
                    "type": "number",
                    "example": 10.3
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-02-22T15:11:19.25616+07:00"
                },
                "id": {
                    "type": "string",
                    "example": "d470a4f0-cd65-497d-9198-c16bbf670447"
                },
                "transaction_type": {
                    "type": "string",
                    "example": "credit"
                }
            }
        },
        "UpdateByIdTransactionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/UpdateTransactionResponse"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "UpdateTransactionRequest": {
            "type": "object",
            "required": [
                "account_number"
            ],
            "properties": {
                "account_number": {
                    "type": "string",
                    "example": "233455011"
                },
                "amount": {
                    "type": "number",
                    "example": 10.3
                },
                "transaction_type": {
                    "type": "string",
                    "example": "credit"
                }
            }
        },
        "UpdateTransactionResponse": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string",
                    "example": "233455011"
                },
                "amount": {
                    "type": "number",
                    "example": 10.3
                },
                "id": {
                    "type": "string",
                    "example": "d470a4f0-cd65-497d-9198-c16bbf670447"
                },
                "transaction_type": {
                    "type": "string",
                    "example": "credit"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-02-22T15:11:19.25616+07:00"
                }
            }
        }
    }
}